# Update generate-docs test to include --help flag