--- a/cmd/minikube/cmd/generate-docs_test.go
+++ b/cmd/minikube/cmd/generate-docs_test.go
@@ -28,7 +28,7 @@ import (
 )
 
 func TestGenerateDocs(t *testing.T) {
-	pflag.BoolP("help", "h", false, "")
+	pflag.BoolP("help", "h", false, "") // avoid 'Docs are not updated. Please run `make generate-docs` to update commands documentation' error
 	dir := "../../../site/content/en/docs/commands/"
 
 	for _, sc := range RootCmd.Commands() {