fixes: #9631 

updates generate-docs test to reflect #9622 and #9614

### before
currently, ci TestGenerateDocs (generate-docs) unit test fails for not including --help flag:
>❯ make clean && make && make test

result:
<details>
<pre>
rm -rf ./out
rm -f pkg/minikube/assets/assets.go
rm -f pkg/minikube/translate/translations.go
rm -rf ./vendor
PATH="/home/prezha/bin:/usr/local/bin:/usr/bin:/bin:/usr/local/go/bin:/usr/local/rar:/home/prezha/bin/rar:/home/prezha/bin/ffmpeg:/home/prezha/dev/go/bin:/home/prezha/dev/protoc/bin:/home/prezha/bin/node/v14/bin:/home/prezha/dev/dotnet/v3:/home/prezha/bin/wkhtmltox/bin:/home/prezha/bin/mongodb:/home/prezha/bin/azure:/home/prezha/bin/k8s:/home/prezha/dev/go/bin" go-bindata -nomemcopy -o pkg/minikube/assets/assets.go -pkg assets deploy/addons/...
gofmt -s -w pkg/minikube/assets/assets.go
PATH="/home/prezha/bin:/usr/local/bin:/usr/bin:/bin:/usr/local/go/bin:/usr/local/rar:/home/prezha/bin/rar:/home/prezha/bin/ffmpeg:/home/prezha/dev/go/bin:/home/prezha/dev/protoc/bin:/home/prezha/bin/node/v14/bin:/home/prezha/dev/dotnet/v3:/home/prezha/bin/wkhtmltox/bin:/home/prezha/bin/mongodb:/home/prezha/bin/azure:/home/prezha/bin/k8s:/home/prezha/dev/go/bin" go-bindata -nomemcopy -o pkg/minikube/translate/translations.go -pkg translate translations/...
gofmt -s -w pkg/minikube/translate/translations.go
go build  -tags "go_getter_nos3 go_getter_nogcs" -ldflags="-X k8s.io/minikube/pkg/version.version=v1.14.2 -X k8s.io/minikube/pkg/version.isoVersion=v1.14.0 -X k8s.io/minikube/pkg/version.isoPath=minikube/iso -X k8s.io/minikube/pkg/version.gitCommitID="18b0c00698dc593b78d677686205e35c3ebb7942" -X k8s.io/minikube/pkg/version.storageProvisionerVersion=v3" -o out/minikube k8s.io/minikube/cmd/minikube
MINIKUBE_LDFLAGS="-X k8s.io/minikube/pkg/version.version=v1.14.2 -X k8s.io/minikube/pkg/version.isoVersion=v1.14.0 -X k8s.io/minikube/pkg/version.isoPath=minikube/iso -X k8s.io/minikube/pkg/version.gitCommitID="18b0c00698dc593b78d677686205e35c3ebb7942" -X k8s.io/minikube/pkg/version.storageProvisionerVersion=v3" ./test.sh
= make lint =============================================================
golangci/golangci-lint info checking GitHub for tag 'v1.30.0'
golangci/golangci-lint info found version: 1.30.0 for v1.30.0/linux/amd64
golangci/golangci-lint info installed out/linters/golangci-lint
ok
= go mod ================================================================
ok
= boilerplate ===========================================================
ok
= schema_check ==========================================================
ok
= go test ===============================================================
--- FAIL: TestGenerateDocs (0.02s)
    --- FAIL: TestGenerateDocs/addons (0.00s)
        generate-docs_test.go:47: Docs are not updated. Please run `make generate-docs` to update commands documentation:   (
                """
                ... // 17 identical lines
                ```
              
            -   ### Options
            +   ### Options inherited from parent commands
              
                ```
            -     -h, --help   help for addons
            -   ```
            - 
            -   ### Options inherited from parent commands
            - 
            -   ```
                      --add_dir_header                   If true, adds the file directory to the header of the log messages
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 22 identical lines
                ```
              
            -   ### Options
            +   ### Options inherited from parent commands
              
                ```
            -     -h, --help   help for configure
            -   ```
            - 
            -   ### Options inherited from parent commands
            - 
            -   ```
                      --add_dir_header                   If true, adds the file directory to the header of the log messages
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 22 identical lines
                ```
              
            -   ### Options
            +   ### Options inherited from parent commands
              
                ```
            -     -h, --help   help for disable
            -   ```
            - 
            -   ### Options inherited from parent commands
            - 
            -   ```
                      --add_dir_header                   If true, adds the file directory to the header of the log messages
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 9 identical lines
                      --vmodule moduleSpec               comma-separated list of pattern=N settings for file-filtered logging
                ```
            - 
            -   ## minikube addons enable
            - 
            -   Enables the addon w/ADDON_NAME within minikube (example: minikube addons enable dashboard). For a list of available addons use: minikube addons list 
              
            -   ### Synopsis
            +   ## minikube addons enable
                ... // 136 identical, 19 removed, and 13 inserted lines
                """
              )
    --- FAIL: TestGenerateDocs/cache (0.00s)
        generate-docs_test.go:47: Docs are not updated. Please run `make generate-docs` to update commands documentation:   (
                """
                ... // 13 identical lines
                Add, delete, or push a local image into minikube
              
            -   ### Options
            +   ### Options inherited from parent commands
              
                ```
            -     -h, --help   help for cache
            -   ```
            - 
            -   ### Options inherited from parent commands
            - 
            -   ```
                      --add_dir_header                   If true, adds the file directory to the header of the log messages
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 22 identical lines
                ```
              
            -   ### Options
            +   ### Options inherited from parent commands
              
                ```
            -     -h, --help   help for add
            -   ```
            - 
            -   ### Options inherited from parent commands
            - 
            -   ```
                      --add_dir_header                   If true, adds the file directory to the header of the log messages
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 22 identical lines
                ```
              
            -   ### Options
            +   ### Options inherited from parent commands
              
                ```
            -     -h, --help   help for delete
            -   ```
            - 
            -   ### Options inherited from parent commands
            - 
            -   ```
                      --add_dir_header                   If true, adds the file directory to the header of the log messages
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 9 identical lines
                      --vmodule moduleSpec               comma-separated list of pattern=N settings for file-filtered logging
                ```
            - 
            -   ## minikube cache help
            - 
            -   Help about any command
              
            -   ### Synopsis
            +   ## minikube cache help
                ... // 94 identical, 19 removed, and 13 inserted lines
                """
              )
    --- FAIL: TestGenerateDocs/completion (0.00s)
        generate-docs_test.go:47: Docs are not updated. Please run `make generate-docs` to update commands documentation:   (
                """
                ... // 36 identical lines
                ```
                minikube completion SHELL [flags]
            -   ```
            - 
            -   ### Options
            - 
            -   ```
            -     -h, --help   help for completion
                ```
              
                ... // 4 identical lines
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 13 identical lines
                """
              )
    --- FAIL: TestGenerateDocs/config (0.00s)
        generate-docs_test.go:47: Docs are not updated. Please run `make generate-docs` to update commands documentation:   (
                """
                ... // 47 identical lines
                ```
              
            -   ### Options
            +   ### Options inherited from parent commands
              
                ```
            -     -h, --help   help for config
            -   ```
            - 
            -   ### Options inherited from parent commands
            - 
            -   ```
                      --add_dir_header                   If true, adds the file directory to the header of the log messages
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 28 identical lines
              
                ```
            -     -h, --help            help for defaults
                      --output string   Output format. Accepted values: [json]
                ```
                ... // 5 identical lines
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 22 identical lines
                ```
              
            -   ### Options
            +   ### Options inherited from parent commands
              
                ```
            -     -h, --help   help for get
            -   ```
            - 
            -   ### Options inherited from parent commands
            - 
            -   ```
                      --add_dir_header                   If true, adds the file directory to the header of the log messages
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 23 identical lines
                ```
              
            -   ### Options
            +   ### Options inherited from parent commands
              
                ```
            -     -h, --help   help for help
            -   ```
            - 
            -   ### Options inherited from parent commands
            - 
            -   ```
                      --add_dir_header                   If true, adds the file directory to the header of the log messages
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 9 identical lines
                      --vmodule moduleSpec               comma-separated list of pattern=N settings for file-filtered logging
                ```
            - 
            -   ## minikube config set
            - 
            -   Sets an individual value in a minikube config file
                ... // 95 identical, 20 removed, and 14 inserted lines
                """
              )
    --- FAIL: TestGenerateDocs/dashboard (0.00s)
        generate-docs_test.go:47: Docs are not updated. Please run `make generate-docs` to update commands documentation:   strings.Join({
                ... // 20 identical lines
                "",
                "```",
            -   "  -h, --help   help for dashboard",
            -   "      --url    Display dashboard URL instead of opening a browser",
            +   "      --url   Display dashboard URL instead of opening a browser",
                "```",
                "",
                ... // 4 identical lines
                "      --alsologtostderr                  log to standard error as well as files",
                `  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")`,
            +   "  -h, --help                             ",
                "      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)",
                "      --log_dir string                   If non-empty, write log files in this directory",
                ... // 13 identical lines
              }, "\n")
    --- FAIL: TestGenerateDocs/delete (0.00s)
        generate-docs_test.go:47: Docs are not updated. Please run `make generate-docs` to update commands documentation:   (
                """
                ... // 22 identical lines
                ```
                      --all     Set flag to delete all profiles
            -     -h, --help    help for delete
                      --purge   Set this flag to delete the '.minikube' folder from your user directory.
                ```
                ... // 5 identical lines
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 13 identical lines
                """
              )
    --- FAIL: TestGenerateDocs/docker-env (0.00s)
        generate-docs_test.go:47: Docs are not updated. Please run `make generate-docs` to update commands documentation:   (
                """
                ... // 20 identical lines
              
                ```
            -     -h, --help           help for docker-env
                      --no-proxy       Add machine IP to NO_PROXY environment variable
                      --shell string   Force environment to be configured for a specified shell: [fish, cmd, powershell, tcsh, bash, zsh], default is auto-detect
                ... // 7 identical lines
                      --alsologtostderr                  log to standard error as well as files
                  -b, --bootstrapper string              The name of the cluster bootstrapper that will set up the Kubernetes cluster. (default "kubeadm")
            +     -h, --help                             
                      --log_backtrace_at traceLocation   when logging hits line file:N, emit a stack trace (default :0)
                      --log_dir string                   If non-empty, write log files in this directory
                ... // 13 identical lines
                """
              )
…